fix some typo; link the reference to the architectures
