+++ official/vision/beta/modeling/backbones/mobilenet.py
@@ -12,128 +12,17 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-"""Contains definitions of EfficientNet Networks."""
+"""Contains definitions of Mobilenet Networks."""
 
-import math
 from typing import Text, Optional, Dict
 
 # Import libraries
-from absl import logging
 import tensorflow as tf
-from official.modeling import tf_utils
 from official.vision.beta.modeling.layers import nn_blocks
 from official.vision.beta.modeling.layers import nn_layers
 
 layers = tf.keras.layers
 regularizers = tf.keras.regularizers
-
-
-class Conv2DBNBlock(tf.keras.layers.Layer):
-  """An convolution block with batch normalization."""
-
-  def __init__(self,
-               filters: int,
-               kernel_size: int = 3,
-               strides: int = 1,
-               use_bias: bool = False,
-               activation: Text = 'relu6',
-               kernel_initializer: Text = 'VarianceScaling',
-               kernel_regularizer: Optional[regularizers.Regularizer] = None,
-               bias_regularizer: Optional[regularizers.Regularizer] = None,
-               use_normalization: bool = True,
-               use_sync_bn: bool = False,
-               norm_momentum: float = 0.99,
-               norm_epsilon: float = 0.001,
-               **kwargs):
-    """An convolution block with batch normalization.
-
-    Args:
-      filters: `int` number of filters for the first two convolutions. Note that
-        the third and final convolution will use 4 times as many filters.
-      kernel_size: `int` an integer specifying the height and width of the
-      2D convolution window.
-      strides: `int` block stride. If greater than 1, this block will ultimately
-        downsample the input.
-      use_bias: if True, use biase in the convolution layer.
-      activation: `str` name of the activation function.
-      kernel_size: `int` kernel_size of the conv layer.
-      kernel_initializer: kernel_initializer for convolutional layers.
-      kernel_regularizer: tf.keras.regularizers.Regularizer object for Conv2D.
-                          Default to None.
-      bias_regularizer: tf.keras.regularizers.Regularizer object for Conv2d.
-                        Default to None.
-      use_normalization: if True, use batch normalization.
-      use_sync_bn: if True, use synchronized batch normalization.
-      norm_momentum: `float` normalization momentum for the moving average.
-      norm_epsilon: `float` small float added to variance to avoid dividing by
-        zero.
-      **kwargs: keyword arguments to be passed.
-    """
-    super(Conv2DBNBlock, self).__init__(**kwargs)
-    self._filters = filters
-    self._kernel_size = kernel_size
-    self._strides = strides
-    self._activation = activation
-    self._use_bias = use_bias
-    self._kernel_initializer = kernel_initializer
-    self._kernel_regularizer = kernel_regularizer
-    self._bias_regularizer = bias_regularizer
-    self._use_normalization = use_normalization
-    self._use_sync_bn = use_sync_bn
-    self._norm_momentum = norm_momentum
-    self._norm_epsilon = norm_epsilon
-
-    if use_sync_bn:
-      self._norm = tf.keras.layers.experimental.SyncBatchNormalization
-    else:
-      self._norm = tf.keras.layers.BatchNormalization
-    if tf.keras.backend.image_data_format() == 'channels_last':
-      self._bn_axis = -1
-    else:
-      self._bn_axis = 1
-    self._activation_fn = tf_utils.get_activation(activation)
-
-  def get_config(self):
-    config = {
-        'filters': self._filters,
-        'strides': self._strides,
-        'kernel_size': self._kernel_size,
-        'use_bias': self._use_bias,
-        'kernel_initializer': self._kernel_initializer,
-        'kernel_regularizer': self._kernel_regularizer,
-        'bias_regularizer': self._bias_regularizer,
-        'activation': self._activation,
-        'use_sync_bn': self._use_sync_bn,
-        'use_normalization': self._use_normalization,
-        'norm_momentum': self._norm_momentum,
-        'norm_epsilon': self._norm_epsilon
-    }
-    base_config = super(Conv2DBNBlock, self).get_config()
-    return dict(list(base_config.items()) + list(config.items()))
-
-  def build(self, input_shape):
-    self._conv0 = tf.keras.layers.Conv2D(
-        filters=self._filters,
-        kernel_size=self._kernel_size,
-        strides=self._strides,
-        padding='same',
-        use_bias=self._use_bias,
-        kernel_initializer=self._kernel_initializer,
-        kernel_regularizer=self._kernel_regularizer,
-        bias_regularizer=self._bias_regularizer)
-    if self._use_normalization:
-      self._norm0 = self._norm(
-          axis=self._bn_axis,
-          momentum=self._norm_momentum,
-          epsilon=self._norm_epsilon)
-
-    super(Conv2DBNBlock, self).build(input_shape)
-
-  def call(self, inputs, training=None):
-    x = self._conv0(inputs)
-    if self._use_normalization:
-      x = self._norm0(x)
-    return self._activation_fn(x)
 
 
 class GlobalPoolingBlock(tf.keras.layers.Layer):
@@ -146,6 +35,14 @@ class GlobalPoolingBlock(tf.keras.layers.Layer):
     return outputs
 
 
+"""
+Architecture: https://arxiv.org/abs/1704.04861.
+
+"MobileNets: Efficient Convolutional Neural Networks for
+  Mobile Vision Applications"
+Andrew G. Howard, Menglong Zhu, Bo Chen, Dmitry Kalenichenko, Weijun Wang,
+  Tobias Weyand, Marco Andreetto, Hartwig Adam
+"""
 MNV1_BLOCK_SPECS = {
     'spec_name': 'MobileNetV1',
     'block_spec_schema': ['block_fn', 'kernel_size', 'strides', 'filters'],
@@ -167,6 +64,12 @@ MNV1_BLOCK_SPECS = {
     ]
 }
 
+"""
+Architecture: https://arxiv.org/abs/1801.04381
+
+"MobileNetV2: Inverted Residuals and Linear Bottlenecks"
+Mark Sandler, Andrew Howard, Menglong Zhu, Andrey Zhmoginov, Liang-Chieh Chen
+"""
 MNV2_BLOCK_SPECS = {
     'spec_name': 'MobileNetV2',
     'block_spec_schema': ['block_fn', 'kernel_size', 'strides', 'filters',
@@ -202,6 +105,13 @@ MNV2_BLOCK_SPECS = {
     ]
 }
 
+"""
+Architecture: https://arxiv.org/abs/1905.02244
+
+"Searching for MobileNetV3"
+Andrew Howard, Mark Sandler, Grace Chu, Liang-Chieh Chen, Bo Chen, Mingxing Tan, 
+Weijun Wang, Yukun Zhu, Ruoming Pang, Vijay Vasudevan, Quoc V. Le, Hartwig Adam
+"""
 MNV3Large_BLOCK_SPECS = {
     'spec_name': 'MobileNetV3Large',
     'block_spec_schema': ['block_fn', 'kernel_size', 'strides', 'filters',
@@ -267,6 +177,10 @@ MNV3Small_BLOCK_SPECS = {
     ]
 }
 
+"""
+The EdgeTPU version is taken from
+github.com/tensorflow/models/blob/master/research/slim/nets/mobilenet/mobilenet_v3.py
+"""
 MNV3EdgeTPU_BLOCK_SPECS = {
     'spec_name': 'MobileNetV3EdgeTPU',
     'block_spec_schema': ['block_fn', 'kernel_size', 'strides', 'filters',
@@ -313,20 +227,19 @@ SUPPORTED_SPECS_MAP = {
     'MobileNetV2': MNV2_BLOCK_SPECS,
     'MobileNetV3Large': MNV3Large_BLOCK_SPECS,
     'MobileNetV3Small': MNV3Small_BLOCK_SPECS,
-    'MobileNetV3EdgeTPU': MNV3EdgeTPU_BLOCK_SPECS
+    'MobileNetV3EdgeTPU': MNV3EdgeTPU_BLOCK_SPECS,
 }
 
 BLOCK_FN_MAP = {
-    'convbn': Conv2DBNBlock,
-    'gpooling': GlobalPoolingBlock,
+    'convbn': nn_blocks.Conv2DBNBlock,
     'depsepconv': nn_blocks.DepthwiseSeparableConvBlock,
     'mbconv': nn_blocks.InvertedBottleneckBlock,
-
+    'gpooling': GlobalPoolingBlock,
 }
 
 
 class BlockSpec(object):
-  """A container class that specifies the block configuration for MnasNet."""
+  """A container class that specifies the block configuration for MobileNet."""
 
   def __init__(self,
                block_fn: Text = 'convbn',
@@ -381,9 +294,6 @@ def block_spec_decoder(specs: Dict,
   spec_name = specs['spec_name']
   block_spec_schema = specs['block_spec_schema']
   block_specs = specs['block_specs']
-
-  if spec_name not in SUPPORTED_SPECS_MAP:
-    raise ValueError('Model spec: {} is not supported !'.format(spec_name))
 
   if len(block_specs) == 0:
     raise ValueError('The block spec cannot be empty for {} !'.format(spec_name))
@@ -441,8 +351,9 @@ class MobileNet(tf.keras.Model):
     """
 
     Args:
-      model_id: `str` version of MobileNet. The supported values are MobileNetV2',
-      'MobileNetV3Large', 'MobileNetV3Small', and 'MobileNetV3EdgeTPU'.
+      model_id: `str` version of MobileNet. The supported values
+        are 'MobileNetV1', 'MobileNetV2', 'MobileNetV3Large', 'MobileNetV3Small',
+        and 'MobileNetV3EdgeTPU'.
       width_multiplier: `float` multiplier for the depth (number of channels)
         for all convolution ops. The value must be greater than zero. Typical
         usage will be to set this value in (0, 1) to reduce the number of
@@ -577,7 +488,7 @@ class MobileNet(tf.keras.Model):
 
       if block_def.block_fn == 'convbn':
 
-        net = Conv2DBNBlock(
+        net = nn_blocks.Conv2DBNBlock(
             filters=block_def.filters,
             kernel_size=block_def.kernel_size,
             strides=block_def.strides,
